Decouple BackgroundIntensity enum from LaunchAgentManager
